+++ src/clients/inworld.client.ts
@@ -18,6 +18,7 @@ import { GrpcAudioPlayback } from '../components/sound/grpc_audio.playback';
 import { GrpcAudioRecorder } from '../components/sound/grpc_audio.recorder';
 import { GrpcWebRtcLoopbackBiDiSession } from '../components/sound/grpc_web_rtc_loopback_bidi.session';
 import { InworldPacket } from '../entities/inworld_packet.entity';
+import { isNaturalNumber } from '../guard/number';
 import { ConnectionService } from '../services/connection.service';
 import { InworldConnectionService } from '../services/inworld_connection.service';
 
@@ -149,9 +150,12 @@ export class InworldClient<
   build() {
     this.validate();
 
+    const config = this.buildConfiguration();
+
     const webRtcLoopbackBiDiSession = new GrpcWebRtcLoopbackBiDiSession();
     const grpcAudioRecorder = new GrpcAudioRecorder();
     const grpcAudioPlayer = new GrpcAudioPlayback<InworldPacketT>({
+      audioPlaybackConfig: config.audioPlayback,
       onAfterPlaying: this.onAfterPlaying,
       onBeforePlaying: this.onBeforePlaying,
       onStopPlaying: this.onStopPlaying,
@@ -159,12 +163,12 @@ export class InworldClient<
     });
 
     const connection = new ConnectionService<InworldPacketT>({
+      config,
       grpcAudioPlayer,
       webRtcLoopbackBiDiSession,
       name: this.scene,
       user: this.user,
       client: this.client,
-      config: this.buildConfiguration(),
       onError: this.onError,
       onReady: this.onReady,
       onMessage: this.onMessage,
@@ -183,10 +187,11 @@ export class InworldClient<
   }
 
   private buildConfiguration(): InternalClientConfiguration {
-    const { connection = {}, capabilities = {} } = this.config;
+    const { connection = {}, capabilities = {}, ...restConfig } = this.config;
     const { gateway } = connection;
 
     return {
+      ...restConfig,
       connection: {
         ...connection,
         gateway: this.ensureGateway(GRPC_HOSTNAME, gateway),
@@ -231,6 +236,20 @@ export class InworldClient<
   private validate() {
     if (!this.scene) {
       throw Error('Scene name is required');
+    }
+
+    const { audioPlayback } = this.config;
+
+    if (audioPlayback) {
+      if (!isNaturalNumber(audioPlayback.stop.duration)) {
+        throw Error(
+          'Stop duration for audio playback should be a natural number',
+        );
+      }
+
+      if (!isNaturalNumber(audioPlayback.stop.ticks)) {
+        throw Error('Stop ticks for audio playback should be a natural number');
+      }
     }
   }
 }