+++ src/common/helpers.ts
@@ -4,3 +4,5 @@ export const protoTimestamp = (date?: Date) =>
 export const isIOSMobile = () => {
   return /iPad|iPhone|iPod/.test(navigator.userAgent);
 };
+
+export const interpolate = (p: number) => 0.5 - Math.cos(p * Math.PI) / 2;